refactor(p2p): remove dead code We've also moved all the code that is marked for deprecation into one place: `p2p_deprecate.go`
